yen_per_kg, from scraper to database

+++ probc2024-sd5.sql
@@ -6,23 +6,22 @@
 -- 生成日時: 2024-11-15 19:08:54
 -- サーバのバージョン： 10.4.32-MariaDB
 -- PHP のバージョン: 8.2.12
-
 SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
 START TRANSACTION;
 SET time_zone = "+00:00";
-
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!40101 SET NAMES utf8mb4 */;
-
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */
+;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */
+;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */
+;
+/*!40101 SET NAMES utf8mb4 */
+;
 --
 -- データベース: `probc2024-sd5`
 --
 
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `admins`
 --
@@ -32,10 +31,8 @@ CREATE TABLE `admins` (
   `name` text NOT NULL,
   `password` text NOT NULL,
   `last_login_date` datetime DEFAULT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `contacts`
 --
@@ -46,10 +43,8 @@ CREATE TABLE `contacts` (
   `form_id` int(11) DEFAULT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `form`
 --
@@ -58,10 +53,8 @@ CREATE TABLE `form` (
   `ID` int(11) NOT NULL,
   `product_array` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_bin NOT NULL CHECK (json_valid(`product_array`)),
   `manpower` int(50) NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `products`
 --
@@ -72,33 +65,70 @@ CREATE TABLE `products` (
   `summary` text DEFAULT NULL,
   `desc` text DEFAULT NULL,
   `product_categories_ID` int(50) NOT NULL,
-  `kg_per_1a` int(50) NOT NULL,
-  `yen_per_1a` int(50) NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+  `kg_per_1a` DOUBLE NOT NULL,
+  `yen_per_1a` DOUBLE NOT NULL
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 --
 -- テーブルのデータのダンプ `products`
 --
 
-INSERT INTO `products` (`ID`, `name`, `summary`, `desc`, `product_categories_ID`, `kg_per_1a`, `yen_per_1kg`) VALUES
-(1, 'あんぽ柿', '品種は平核無柿、蜂屋柿など', NULL, 1, 78, 240),
-(2, 'いちご', '品種はとちおとめ、さちのか、など', NULL, 2, 218, 1190),
-(3, 'いんげん', '品種はいちず、鴨川グリーンなど', NULL, 2, 72, 8678),
-(4, 'きゅうり', '品種はアンコール10、パイロット2号、南極1号など', NULL, 2, 580, 233),
-(5, 'さくらんぼ', '品種は佐藤錦など', NULL, 1, 2835, 40),
-(6, 'さやえんどう', '品種は姫みどり、ゆうさやなど', NULL, 2, 43, 1157),
-(7, 'しいたけ', NULL, NULL, 3, 822, 928),
-(8, '春菊', NULL, NULL, 2, 119, 516),
-(9, '西洋なし', '品種はラ・フランス、ル・レクチェ、ゼネラル・レクラークなど', NULL, 1, 157, 373),
-(10, 'ニラ', '品種はワンダーグリーン、パワフルグリーンベルトなど', NULL, 2, 168, 546),
-(11, '花わさび', NULL, NULL, 2, 1, 6000),
-(12, 'ピーマン', NULL, NULL, 2, 346, 346),
-(13, 'ぶどう', '品種は巨峰、高尾など', NULL, 1, 94, 1062),
-(14, '桃', '品種はあかつき、川中島白桃、ゆうぞらなど', NULL, 1, 158, 695),
-(15, 'りんご', '品種は王林、ふじなど', NULL, 1, 159, 287);
-
+INSERT INTO `products` (
+    `ID`,
+    `name`,
+    `summary`,
+    `desc`,
+    `product_categories_ID`,
+    `kg_per_1a`,
+    `yen_per_1kg`
+  )
+VALUES (1, 'あんぽ柿', '品種は平核無柿、蜂屋柿など', NULL, 1, 78, 240),
+  (2, 'いちご', '品種はとちおとめ、さちのか、など', NULL, 2, 218, 1190),
+  (3, 'いんげん', '品種はいちず、鴨川グリーンなど', NULL, 2, 72, 8678),
+  (
+    4,
+    'きゅうり',
+    '品種はアンコール10、パイロット2号、南極1号など',
+    NULL,
+    2,
+    580,
+    233
+  ),
+  (5, 'さくらんぼ', '品種は佐藤錦など', NULL, 1, 2835, 40),
+  (6, 'さやえんどう', '品種は姫みどり、ゆうさやなど', NULL, 2, 43, 1157),
+  (7, 'しいたけ', NULL, NULL, 3, 822, 928),
+  (8, '春菊', NULL, NULL, 2, 119, 516),
+  (
+    9,
+    '西洋なし',
+    '品種はラ・フランス、ル・レクチェ、ゼネラル・レクラークなど',
+    NULL,
+    1,
+    157,
+    373
+  ),
+  (
+    10,
+    'ニラ',
+    '品種はワンダーグリーン、パワフルグリーンベルトなど',
+    NULL,
+    2,
+    168,
+    546
+  ),
+  (11, '花わさび', NULL, NULL, 2, 1, 6000),
+  (12, 'ピーマン', NULL, NULL, 2, 346, 346),
+  (13, 'ぶどう', '品種は巨峰、高尾など', NULL, 1, 94, 1062),
+  (
+    14,
+    '桃',
+    '品種はあかつき、川中島白桃、ゆうぞらなど',
+    NULL,
+    1,
+    158,
+    695
+  ),
+  (15, 'りんご', '品種は王林、ふじなど', NULL, 1, 159, 287);
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `products_relations`
 --
@@ -108,10 +138,8 @@ CREATE TABLE `products_relations` (
   `product1_id` int(50) NOT NULL,
   `product2_id` int(50) NOT NULL,
   `rating` int(50) NOT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 -- --------------------------------------------------------
-
 --
 -- テーブルの構造 `product_categories`
 --
@@ -120,17 +148,15 @@ CREATE TABLE `product_categories` (
   `ID` int(11) NOT NULL,
   `name` text NOT NULL,
   `summary` text DEFAULT NULL
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
-
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4 COLLATE = utf8mb4_general_ci;
 --
 -- テーブルのデータのダンプ `product_categories`
 --
 
-INSERT INTO `product_categories` (`ID`, `name`, `summary`) VALUES
-(1, '果樹類', NULL),
-(2, '野菜類', NULL),
-(3, 'きのこ類', NULL);
-
+INSERT INTO `product_categories` (`ID`, `name`, `summary`)
+VALUES (1, '果樹類', NULL),
+  (2, '野菜類', NULL),
+  (3, 'きのこ類', NULL);
 --
 -- ダンプしたテーブルのインデックス
 --
@@ -139,42 +165,36 @@ INSERT INTO `product_categories` (`ID`, `name`, `summary`) VALUES
 -- テーブルのインデックス `admins`
 --
 ALTER TABLE `admins`
-  ADD PRIMARY KEY (`ID`);
-
+ADD PRIMARY KEY (`ID`);
 --
 -- テーブルのインデックス `contacts`
 --
 ALTER TABLE `contacts`
-  ADD PRIMARY KEY (`ID`),
+ADD PRIMARY KEY (`ID`),
   ADD KEY `form_id` (`form_id`);
-
 --
 -- テーブルのインデックス `form`
 --
 ALTER TABLE `form`
-  ADD PRIMARY KEY (`ID`);
-
+ADD PRIMARY KEY (`ID`);
 --
 -- テーブルのインデックス `products`
 --
 ALTER TABLE `products`
-  ADD PRIMARY KEY (`ID`),
+ADD PRIMARY KEY (`ID`),
   ADD KEY `products_ibfk_1` (`product_categories_ID`);
-
 --
 -- テーブルのインデックス `products_relations`
 --
 ALTER TABLE `products_relations`
-  ADD PRIMARY KEY (`ID`),
+ADD PRIMARY KEY (`ID`),
   ADD KEY `product1_id` (`product1_id`),
   ADD KEY `product2_id` (`product2_id`);
-
 --
 -- テーブルのインデックス `product_categories`
 --
 ALTER TABLE `product_categories`
-  ADD PRIMARY KEY (`ID`);
-
+ADD PRIMARY KEY (`ID`);
 --
 -- ダンプしたテーブルの AUTO_INCREMENT
 --
@@ -183,38 +203,34 @@ ALTER TABLE `product_categories`
 -- テーブルの AUTO_INCREMENT `admins`
 --
 ALTER TABLE `admins`
-  MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
-
+MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
 --
 -- テーブルの AUTO_INCREMENT `contacts`
 --
 ALTER TABLE `contacts`
-  MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
-
+MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
 --
 -- テーブルの AUTO_INCREMENT `form`
 --
 ALTER TABLE `form`
-  MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
-
+MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT;
 --
 -- テーブルの AUTO_INCREMENT `products`
 --
 ALTER TABLE `products`
-  MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=16;
-
+MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT,
+  AUTO_INCREMENT = 16;
 --
 -- テーブルの AUTO_INCREMENT `products_relations`
 --
 ALTER TABLE `products_relations`
-  MODIFY `ID` int(50) NOT NULL AUTO_INCREMENT;
-
+MODIFY `ID` int(50) NOT NULL AUTO_INCREMENT;
 --
 -- テーブルの AUTO_INCREMENT `product_categories`
 --
 ALTER TABLE `product_categories`
-  MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=4;
-
+MODIFY `ID` int(11) NOT NULL AUTO_INCREMENT,
+  AUTO_INCREMENT = 4;
 --
 -- ダンプしたテーブルの制約
 --
@@ -223,22 +239,22 @@ ALTER TABLE `product_categories`
 -- テーブルの制約 `contacts`
 --
 ALTER TABLE `contacts`
-  ADD CONSTRAINT `contacts_ibfk_1` FOREIGN KEY (`form_id`) REFERENCES `form` (`ID`);
-
+ADD CONSTRAINT `contacts_ibfk_1` FOREIGN KEY (`form_id`) REFERENCES `form` (`ID`);
 --
 -- テーブルの制約 `products`
 --
 ALTER TABLE `products`
-  ADD CONSTRAINT `products_ibfk_1` FOREIGN KEY (`product_categories_ID`) REFERENCES `product_categories` (`ID`);
-
+ADD CONSTRAINT `products_ibfk_1` FOREIGN KEY (`product_categories_ID`) REFERENCES `product_categories` (`ID`);
 --
 -- テーブルの制約 `products_relations`
 --
 ALTER TABLE `products_relations`
-  ADD CONSTRAINT `products_relations_ibfk_1` FOREIGN KEY (`product1_id`) REFERENCES `products` (`ID`),
+ADD CONSTRAINT `products_relations_ibfk_1` FOREIGN KEY (`product1_id`) REFERENCES `products` (`ID`),
   ADD CONSTRAINT `products_relations_ibfk_2` FOREIGN KEY (`product2_id`) REFERENCES `products` (`ID`);
 COMMIT;
-
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */
+;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */
+;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */
+;
